src/test/java/com/jack/jpastd/service/MemberServiceTest.java
package com.jack.jpastd.service;

import com.jack.jpastd.domain.Member;
import com.jack.jpastd.repository.MemberRepository;

import static org.junit.jupiter.api.Assertions.*;

import javax.persistence.EntityManager;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.TestPropertySource;
import org.springframework.transaction.annotation.Transactional;

@SpringBootTest
@Transactional
@ActiveProfiles("test")
public class MemberServiceTest {

    @Autowired MemberService memberService;
    @Autowired MemberRepository memberRepository;
    @Autowired EntityManager em;

    @Test
    //@Rollback(false)
    public void 회원가입() throws Exception {
        // given
        Member member = new Member();
        member.setName("kim");
        
        // when
        Long saveId = memberService.join(member);

        // then 
        assertEquals(member, memberRepository.findOne(saveId));
    }

    @Test
    public void 중복_회원_예외() throws Exception {
        // given
        Member member1 = new Member();
        member1.setName("kim");

        Member member2 = new Member();
        member2.setName("kim");
        
        // when
        memberService.join(member1);
        //memberService.join(member2);

        // then 
       IllegalStateException thrown = assertThrows(IllegalStateException.class, () -> memberService.join(member2));
       assertEquals("이미 존재하는 회원입니다.", thrown.getMessage());
    }


}
